+++ NEWS.md
@@ -4,7 +4,10 @@ This page describes the most important changes in `TypeUtils`. The format is bas
 [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 [Semantic Versioning](https://semver.org).
 
-## Unreleased
+## Version 1.9.0 (2025-06-20)
+
+This new minor version adds methods to deal with the numerical precision of numbers and
+objects.
 
 ## Added